fixes #6431 // make sure config file is deleted only during spamfilter_user event.

See merge request dimitri/ispconfig3!1

+++ server/plugins-available/rspamd_plugin.inc.php
@@ -281,7 +281,7 @@ class rspamd_plugin {
 				$app->system->mkdirpath($this->users_config_dir);
 			}
 
-			if((!$this->isValidEmail($app->functions->idn_encode($email_address))) || intval($data['new']['policy_id']) == 0) {
+			if((!$this->isValidEmail($app->functions->idn_encode($email_address))) || intval($data['new']['policy_id']) == 0 && $type == 'spamfilter_user') {
 				if(is_file($settings_file)) {
 					unlink($settings_file);
 				}
